Kovri: client namespace refactor + lib organization
- All client code only requires kovri::client namespace
- Moved code to more appropriate libclient directories
- Specified header locations in relation to libraries

Note: this work will prepare libclient for further work with larger
refactoring and library improvements.

This commit is malleable and the work will most likely be subject to
further revision.

References #339 #98
