+++ src/app/win32_service.cc
@@ -34,13 +34,14 @@
 #define _CRT_SECURE_NO_WARNINGS  // to use freopen
 #endif
 
-#include "win32_service.h"
+#include "app/win32_service.h"
 
 #include <assert.h>
 
 #include <memory>
 
-#include "daemon.h"
+#include "app/daemon.h"
+
 #include "core/util/log.h"
 
 I2PService *I2PService::m_Service = NULL;
