+++ src/client/CMakeLists.txt
@@ -1,17 +1,17 @@
 set(CLIENT_SRC
   "address_book.cc"
   "address_book_storage.cc"
-  "client_context.cc"
-  "datagram.cc"
+  "api/datagram.cc"
+  "api/i2p_control.cc"
+  "api/i2p_control_server.cc"
+  "api/streaming.cc"
+  "context.cc"
   "destination.cc"
-  "i2p_control/i2p_control.cc"
-  "i2p_control/i2p_control_server.cc"
-  "i2p_service.cc"
-  "i2p_tunnel/http_proxy.cc"
-  "i2p_tunnel/i2p_tunnel.cc"
-  "i2p_tunnel/socks.cc"
+  "proxy/http.cc"
+  "proxy/socks.cc"
   "reseed.cc"
-  "streaming.cc"
+  "service.cc"
+  "tunnel.cc"
   "util/http.cc"
   "util/zip.cc")
